+++ srv/backend/lib/Wordwank/Schema/Result/Player.pm
@@ -38,6 +38,11 @@ __PACKAGE__->add_columns(
         data_type => 'text',
         is_nullable => 0,
         default_value => 'en',
+    },
+    lifetime_score => {
+        data_type => 'integer',
+        is_nullable => 0,
+        default_value => 0,
     }
 );